color-analysis/api/requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
google-auth==2.36.0
google-auth-oauthlib==0.4.6
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.21.5
oauthlib==3.2.2
open-clip-torch==2.11.1
packaging==24.2
pandas==1.5.2
parso==0.8.4
pexpect==4.9.0
Pillow==9.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==3.20.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.10.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.11
torch==1.12.1
torch-summary==1.4.5
torchvision==0.13.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.27.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
zipp==3.21.